backend/pyproject.toml
[tool.poetry]
name = "newsmvp"
version = "0.1.0"
description = ""
authors = ["Yura <yura.savaryn.777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.124.4"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
python-dotenv = "^1.2.1"
pandas = "^2.3.3"
requests = "^2.32.5"
feedparser = "^6.0.12"
beautifulsoup4 = "^4.14.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.45"}
asyncpg = "^0.31.0"
pydantic-settings = "^2.12.0"
redis = "^7.1.0"
celery = "^5.6.0"
alembic = "^1.17.2"


[tool.poetry.group.dev.dependencies]
black = "^25.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
